requirements.txt
assertpy
bolt-ta
coverage
nose
